requirements.txt
asgiref==3.8.1
Django==5.1.3
django-cors-headers==4.6.0
django-extensions==3.2.3
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
middleware==1.2.3
mysqlclient==2.2.6
PyJWT==2.10.0
python-dotenv==1.0.1
pytz==2025.1
sqlparse==0.5.2
tzdata==2024.2
